fix(ci): share workspace rbac contract with server build
